+++ src/components/componentBarLineChart/index.ts
@@ -1,0 +1,4 @@
+import ComponentBarLineChart from './componentBarLineChart';
+import { register } from '../../shared/functions';
+
+register('component-bar-line-chart', ComponentBarLineChart);