+++ src/components/pageUser/index.ts
@@ -1,5 +1,6 @@
 import PageUser from './pageUser';
 import { register } from '../../shared/functions';
 import '../componentLineChart/index';
+import '../componentBarLineChart/index';
 
 register('page-user', PageUser);