+++ src/components/componentUserStatus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-user-status
 
+## 0.0.5
+
+### Patch Changes
+
+- Displaying weekly statistics in user page
+
 ## 0.0.4
 
 ### Patch Changes